+++ doc/README.hbs
@@ -14,3 +14,29 @@
 [![downloads](http://img.shields.io/npm/dm/kronos-test-interceptor.svg?style=flat-square)](https://npmjs.org/package/kronos-test-interceptor)
 [![Commitizen friendly](https://img.shields.io/badge/commitizen-friendly-brightgreen.svg)](http://commitizen.github.io/cz-cli/)
 
+kronos-test-interceptor
+=====
+test support for kronos interceptors
+
+# API Reference
+{{#modules~}}
+- {{name}}
+{{/modules}}
+
+{{#functions}}* {{>docs}}
+{{/functions}}
+* * *
+
+install
+=======
+
+With [npm](http://npmjs.org) do:
+
+```shell
+npm install kronos-test-interceptor
+```
+
+license
+=======
+
+BSD-2-Clause
