--- index.js
@@ -1,9 +1,0 @@
-/* jslint node: true, esnext: true */
-'use strict';
-
-exports.MockReceiveInterceptor = require('./lib/mockReceiveInterceptor').MockReceiveInterceptor;
-exports.mochaInterceptorTest = require('./lib/mochaInterceptorTest').mochaInterceptorTest;
-exports.testResponseHandler = require('./lib/mochaInterceptorTest').testResponseHandler;
-exports.TestInterceptor = require('./lib/mochaInterceptorTest').TestInterceptor;
-
-exports.mockReadStreamFactory = require('./lib/mockReadStream');
