+++ rollup.config.js
@@ -3,5 +3,6 @@
 
 export default {
   format: 'cjs',
-  plugins: []
+  plugins: [],
+  external: ['kronos-interceptor', 'loglevel-mixin']
 };
