+++ src/module.js
@@ -1,0 +1,27 @@
+/* jslint node: true, esnext: true */
+'use strict';
+
+import {
+  MockReceiveInterceptor
+}
+from './mockReceiveInterceptor';
+
+import {
+  TestInterceptor,
+  testResponseHandler,
+  mochaInterceptorTest
+}
+from './mochaInterceptorTest';
+
+import {
+  mockReadStreamFactory
+} 
+from './mockReadStream';
+
+export {
+  MockReceiveInterceptor,
+  TestInterceptor,
+  testResponseHandler,
+  mochaInterceptorTest,
+  mockReadStreamFactory
+};
